Fix warning about compare signed and unsigned.

+++ cocos/platform/CCFileUtils.h
@@ -90,7 +90,7 @@ class ResizableBufferAdapter<Data> : public ResizableBuffer {
 public:
     explicit ResizableBufferAdapter(BufferType* buffer) : _buffer(buffer) {}
     virtual void resize(size_t size) override {
-        if (_buffer->getSize() < size) {
+        if (static_cast<size_t>(_buffer->getSize()) < size) {
             auto old = _buffer->getBytes();
             void* buffer = realloc(old, size);
             if (buffer)
